Cast to long for c++11.
git-svn-id: http://svn.boost.org/svn/boost/trunk@86693 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ libs/mpl/test/bitwise.cpp
@@ -37,9 +37,9 @@ MPL_TEST_CASE()
     MPL_ASSERT_RELATION( (bitor_<_0,_0>::value), ==, 0 );
     MPL_ASSERT_RELATION( (bitor_<_1,_0>::value), ==, 1 );
     MPL_ASSERT_RELATION( (bitor_<_0,_1>::value), ==, 1 );
-    MPL_ASSERT_RELATION( (bitor_<_0,_ffffffff>::value), ==, 0xffffffff );
-    MPL_ASSERT_RELATION( (bitor_<_1,_ffffffff>::value), ==, 0xffffffff );
-    MPL_ASSERT_RELATION( (bitor_<_8,_ffffffff>::value), ==, 0xffffffff );
+    MPL_ASSERT_RELATION( static_cast<long>(bitor_<_0,_ffffffff>::value), ==, static_cast<long>(0xffffffff) );
+    MPL_ASSERT_RELATION( static_cast<long>(bitor_<_1,_ffffffff>::value), ==, static_cast<long>(0xffffffff) );
+    MPL_ASSERT_RELATION( static_cast<long>(bitor_<_8,_ffffffff>::value), ==, static_cast<long>(0xffffffff) );
 }
 
 MPL_TEST_CASE()
@@ -47,9 +47,9 @@ MPL_TEST_CASE()
     MPL_ASSERT_RELATION( (bitxor_<_0,_0>::value), ==, 0 );
     MPL_ASSERT_RELATION( (bitxor_<_1,_0>::value), ==, 1 );
     MPL_ASSERT_RELATION( (bitxor_<_0,_1>::value), ==, 1 );
-    MPL_ASSERT_RELATION( (bitxor_<_0,_ffffffff>::value), ==, (0xffffffff ^ 0) );
-    MPL_ASSERT_RELATION( (bitxor_<_1,_ffffffff>::value), ==, (0xffffffff ^ 1) );
-    MPL_ASSERT_RELATION( (bitxor_<_8,_ffffffff>::value), ==, (0xffffffff ^ 8) );
+    MPL_ASSERT_RELATION( static_cast<long>(bitxor_<_0,_ffffffff>::value), ==, static_cast<long>(0xffffffff ^ 0) );
+    MPL_ASSERT_RELATION( static_cast<long>(bitxor_<_1,_ffffffff>::value), ==, static_cast<long>(0xffffffff ^ 1) );
+    MPL_ASSERT_RELATION( static_cast<long>(bitxor_<_8,_ffffffff>::value), ==, static_cast<long>(0xffffffff ^ 8) );
 }
 
 MPL_TEST_CASE()
